# server/docs/swagger.yaml
basePath: /
definitions:
...
  models.AuthRedirectResponse:
    properties:
      location:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.PollCountsResponse:
    properties:
      counts:
        additionalProperties:
          format: int64
          type: integer
        type: object
    type: object
  models.PollOptionRenameRequest:
    properties:
      new_option:
        type: string
      old_option:
        type: string
    type: object
  models.PollOptionRequest:
    properties:
      option:
        type: string
    type: object
  models.PollOptionsResponse:
    properties:
      options:
        items:
          type: string
        type: array
    type: object
  models.PollTitleRequest:
    properties:
      title:
        type: string
    type: object
  models.PollTitleResponse:
    properties:
      title:
        type: string
    type: object
...
  models.SiteSettingsPatchRequest:
    properties:
      site_title:
        type: string
    type: object
  models.SiteSettingsResponse:
    properties:
      site_title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API for The Triangle newspaper CMS
  title: Triangle CMS API
  version: "1.0"
paths:
...
  /v1/auth/callback:
    get:
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/models.AuthRedirectResponse'
      summary: Complete OIDC login callback
      tags:
      - auth
...
  /v1/poll:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PollCountsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get poll counts
      tags:
      - poll
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      parameters:
      - description: Poll vote payload
        in: body
        name: body
        schema:
          $ref: '#/definitions/models.PollOptionRequest'
      - description: Poll vote option (form payload)
        in: formData
        name: poll
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PollCountsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Submit poll vote
      tags:
      - poll
  /v1/poll/options:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Poll option
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PollOptionRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete poll option
      tags:
      - poll
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PollOptionsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get poll options
      tags:
      - poll
    patch:
      consumes:
      - application/json
      parameters:
      - description: Poll option rename
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PollOptionRenameRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Rename poll option
      tags:
      - poll
    post:
      consumes:
      - application/json
      parameters:
      - description: Poll option
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PollOptionRequest'
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add poll option
      tags:
      - poll
  /v1/poll/title:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PollTitleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get poll title
      tags:
      - poll
    patch:
      consumes:
      - application/json
      parameters:
      - description: Poll title update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.PollTitleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PollTitleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update poll title
      tags:
      - poll
...
  /v1/settings/site:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SiteSettingsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get site settings
      tags:
      - settings
    patch:
      consumes:
      - application/json
      parameters:
      - description: Site settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SiteSettingsPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SiteSettingsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update site settings
      tags:
      - settings
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <your-token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...